backend/requirements.txt
flask>=2.1
requests>=2.28
gunicorn>=20.1
python-dotenv>=1.0.0
google-genai
protobuf
google-auth
